chore: move App file to pages folder
