+++ react-development-proj/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import App from './App';
+import App from './pages/App/App';
 
 const router = createBrowserRouter([
   {
