+++ react-development-proj/src/pages/App/App.tsx
@@ -1,11 +1,11 @@
 import './App.css'
-import TitleLayout from './layout/TitleLayout'
+import TitleLayout from '../../layout/TitleLayout'
 
 function App() {
   return (
     <>
       <TitleLayout>
-        {/* Move these in-line styles to a CSS file */}
+        {/* Move these in-line styles to a CSS file? Could get messy in the future */}
         <h1 style={{ fontSize: "6.5em", textAlign: "center" }}>Title</h1>
         <div style={{ textAlign: "center"}}>
           <button className='TitleButtons'>
